refactor: City, Shuttler 서비스 분리
