+++ src/main/java/in/koreatech/koin/domain/bus/service/ShuttleBusService.java
@@ -1,5 +1,6 @@
 package in.koreatech.koin.domain.bus.service;
 
+import java.time.Clock;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -7,7 +8,12 @@ import org.springframework.transaction.annotation.Transactional;
 
 import in.koreatech.koin.domain.bus.dto.ShuttleBusRoutesResponse;
 import in.koreatech.koin.domain.bus.dto.ShuttleBusTimetableResponse;
+import in.koreatech.koin.domain.bus.model.BusRemainTime;
+import in.koreatech.koin.domain.bus.model.enums.BusStation;
+import in.koreatech.koin.domain.bus.model.enums.BusType;
+import in.koreatech.koin.domain.bus.model.mongo.BusCourse;
 import in.koreatech.koin.domain.bus.model.mongo.ShuttleBusRoute;
+import in.koreatech.koin.domain.bus.repository.BusRepository;
 import in.koreatech.koin.domain.bus.repository.ShuttleBusRepository;
 import in.koreatech.koin.domain.version.dto.VersionMessageResponse;
 import in.koreatech.koin.domain.version.service.VersionService;
@@ -20,6 +26,8 @@ public class ShuttleBusService {
 
     private final ShuttleBusRepository shuttleBusRepository;
     private final VersionService versionService;
+    private final BusRepository busRepository;
+    private final Clock clock;
 
     public ShuttleBusRoutesResponse getShuttleBusRoutes() {
         VersionMessageResponse version = versionService.getVersionWithMessage("shuttle_bus_timetable");
@@ -30,5 +38,21 @@ public class ShuttleBusService {
     public ShuttleBusTimetableResponse getShuttleBusTimetable(String id) {
         ShuttleBusRoute shuttleBusRoute = shuttleBusRepository.getById(id);
         return ShuttleBusTimetableResponse.from(shuttleBusRoute);
+    }
+
+    public List<BusRemainTime> getShuttleBusRemainTimes(BusType busType, BusStation depart, BusStation arrival) {
+        List<BusCourse> busCourses = busRepository.findByBusType(busType.getName());
+
+        return busCourses.stream()
+            .map(BusCourse::getRoutes)
+            .flatMap(routes ->
+                routes.stream()
+                    .filter(route -> route.isRunning(clock))
+                    .filter(route -> route.isCorrectRoute(depart, arrival, clock))
+                    .map(route -> route.getRemainTime(depart))
+            )
+            .distinct()
+            .sorted()
+            .toList();
     }
 }
